Refactor: moved from component -> layout
