+++ client/src/routes/NavbarRoute.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { Navbar } from "../components/Navbar";
+import { Navbar } from "../layouts/Navbar";
 
 
 const NavbarRoute = () => {
